+++ anvil_consortium_manager/anvil_api.py
@@ -107,6 +107,17 @@ class AnVILAPIClient:
         url = self.rawls_entry_point + "/api/billing/v2/" + billing_project
         return self.auth_session.get(url, 200)
 
+    def get_billing_projects(self):
+        """Get a list of available billing projects.
+
+        Calls the Sam /api/billing/v2 GET method.
+
+        Returns:
+            requests.Response
+        """
+        url = self.rawls_entry_point + "/api/billing/v2"
+        return self.auth_session.get(url, 200)
+
     def get_groups(self):
         """Get a list of groups that the authenticated account is part of.
 
@@ -403,6 +414,23 @@ class AnVILAPIClient:
         # False here means do not invite unregistered users.
         updates = json.dumps(acl_updates)
         return self.auth_session.patch(url, 200, headers={"Content-type": "application/json"}, data=updates)
+
+    def update_workspace_requester_pays(self, workspace_namespace, workspace_name, requester_pays):
+        """Update the requester pays setting for a workspace.
+        You must be an owner of the workspace to use this method.
+
+        Calls the Rawls /api/workspaces/v2/{workspace_namespace}/{workspace_name}/settings PUT method.
+
+        Args:
+            workspace_namespace (str): The namespace (or billing project) of the workspace.
+            workspace_name (str): The name of the workspace.
+            requester_pays (bool): Whether to enable requester pays for the workspace.
+        Returns:
+            requests.Response
+        """
+        url = self.rawls_entry_point + "/api/workspaces/v2/{}/{}/settings".format(workspace_namespace, workspace_name)
+        setting = [{"config": {"enabled": requester_pays}, "settingType": "GcpBucketRequesterPays"}]
+        return self.auth_session.put(url, 200, headers={"Content-type": "application/json"}, data=json.dumps(setting))
 
 
 class AnVILAPISession(AuthorizedSession):