+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change log
 
-## 0.31.0 (Devel)
+## 0.33.0 (Devel)
 
+* Implement new `anvil_api` method `get_billing_projects` to get a list of billing projects for the user
+* Update `BillingProjectImport` to use the billing project list to provide a dropdown of choices available to import
 * Allow the app to manage workspaces where it is an owner but does not have access (e.g., due to auth domain membership).
 * Rework methods for determining access to workspaces.
     * Add `Workspace.has_account_in_authorization_domain()`, `Workspace.is_shared_with_account()`, `Workspace.is_accessible_by_account()` methods to determine access to a workspace for an account.
@@ -10,6 +12,16 @@
     * Remove old `ManagedGroup` methods used for determining workspace access: `is_in_authorization_domain`, `is_shared`, `has_access`.
     * Remove old `Account` methods used for determine workspace access: `get_accessible_workspaces`, `has_workspace_access`.
 
+## 0.32.0 (2025-07-01)
+
+* Implement audit result caching. Users should set the `ANVIL_AUDIT_CACHE` setting and set an appropriate backend.
+    * Audit views have been split into "run" and "review" views. The "run" view runs the audit and caches the results, while the "review" view displays the cached results.
+    * The `run_anvil_audit` management command can optionally cache the results if called with the `--cache-results` flag.
+* Typo fixes
+
+## 0.31.0 (2025-06-03)
+
+* Add a new to update workspace requester pays status on AnVIL. Note that this removes the "is_requester_pays" field from the `WorkspaceForm`, since updates are now handled via the new `WorkspaceUpdateRequesterPays` view.
 
 ## 0.30.1 (2025-03-07)